pothole/requirements.txt
streamlit==1.32.2
inference-sdk==1.6.0
pillow==10.3.0
requests==2.31.0